# api/cmd/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {},
        "version": "v0.1-beta",
        "title": "Story Protocol API Reference"
    },
    "paths": {
        "/api/v1/assets": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Retrieve a paginated, filtered list of IPAssets",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "IPAssets"
                ],
                "summary": "List IPAssets",
                "parameters": [
                    {
                        "type": "string",
                        "description": "API Key",
                        "name": "X-API-Key",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Query Parameters (",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/betav0.IpAssetRequestBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/betav0.IPAssetsResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/assets/{assetId}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Retrieve an IPAsset",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "IPAssets"
                ],
                "summary": "Get an IPAsset",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Asset ID",
                        "name": "assetId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "API Key",
                        "name": "X-API-Key",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/betav0.IPAssetResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/collections": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Retrieve a paginated, filtered list of Collections",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Collections"
                ],
                "summary": "List Collections",
                "parameters": [
                    {
                        "type": "string",
                        "description": "API Key",
                        "name": "X-API-Key",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Query Parameters (",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/betav0.CollectionsRequestBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/betav0.CollectionsResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/collections/{collectionId}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Retrieve a Collection",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Collections"
                ],
                "summary": "Get an Collection",
                "parameters": [
                    {
                        "type": "string",
                        "description": "API Key",
                        "name": "X-API-Key",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Collection ID",
                        "name": "collectionId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/betav0.CollectionResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/disputes": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Retrieve a paginated, filtered list of Disputes",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Disputes"
                ],
                "summary": "List Disputes",
                "parameters": [
                    {
                        "type": "string",
                        "description": "API Key",
                        "name": "X-API-Key",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Query Parameters (",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/betav0.DisputeRequestBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/betav0.DisputesResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/disputes/{disputeId}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Retrieve a Dispute",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Disputes"
                ],
                "summary": "Get a Dispute",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Dispute ID",
                        "name": "disputeId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "API Key",
                        "name": "X-API-Key",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/betav0.DisputeResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "betav0.Collection": {
            "type": "object",
            "properties": {
                "assetCount": {
                    "type": "string"
                },
                "blockNumber": {
                    "type": "string"
                },
                "blockTimestamp": {
                    "type": "string"
                },
                "cancelledDisputeCount": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "judgedDisputeCount": {
                    "type": "string"
                },
                "licensesCount": {
                    "type": "string"
                },
                "raisedDisputeCount": {
                    "type": "string"
                },
                "resolvedDisputeCount": {
                    "type": "string"
                }
            }
        },
        "betav0.CollectionQueryOptions": {
            "type": "object",
            "properties": {
                "orderBy": {
                    "type": "string"
                },
                "orderDirection": {
                    "type": "string"
                },
                "pagination": {
                    "type": "object",
                    "properties": {
                        "limit": {
                            "type": "integer"
                        },
                        "offset": {
                            "type": "integer"
                        }
                    }
                },
                "where": {
                    "type": "object"
                }
            }
        },
        "betav0.CollectionResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/betav0.Collection"
                }
            }
        },
        "betav0.CollectionsRequestBody": {
            "type": "object",
            "properties": {
                "options": {
                    "$ref": "#/definitions/betav0.CollectionQueryOptions"
                }
            }
        },
        "betav0.CollectionsResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/betav0.Collection"
                    }
                }
            }
        },
        "betav0.Dispute": {
            "type": "object",
            "properties": {
                "arbitrationPolicy": {
                    "type": "string"
                },
                "blockNumber": {
                    "type": "string"
                },
                "blockTimestamp": {
                    "type": "string"
                },
                "currentTag": {
                    "type": "string"
                },
                "data": {
                    "type": "string"
                },
                "evidenceLink": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "initiator": {
                    "type": "string"
                },
                "targetIpId": {
                    "type": "string"
                },
                "targetTag": {
                    "type": "string"
                }
            }
        },
        "betav0.DisputeQueryOptions": {
            "type": "object",
            "properties": {
                "orderBy": {
                    "type": "string"
                },
                "orderDirection": {
                    "type": "string"
                },
                "pagination": {
                    "type": "object",
                    "properties": {
                        "limit": {
                            "type": "integer"
                        },
                        "offset": {
                            "type": "integer"
                        }
                    }
                },
                "where": {
                    "type": "object",
                    "properties": {
                        "currentTag": {
                            "type": "string"
                        },
                        "initiator": {
                            "type": "string"
                        },
                        "targetIpId": {
                            "type": "string"
                        },
                        "targetTag": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "betav0.DisputeRequestBody": {
            "type": "object",
            "properties": {
                "options": {
                    "$ref": "#/definitions/betav0.DisputeQueryOptions"
                }
            }
        },
        "betav0.DisputeResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/betav0.Dispute"
                }
            }
        },
        "betav0.DisputesResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/betav0.Dispute"
                    }
                }
            }
        },
...
        "betav0.IPAsset": {
            "type": "object",
            "properties": {
                "blockNumber": {
                    "type": "string"
                },
                "blockTimestamp": {
                    "type": "string"
                },
                "chainId": {
                    "type": "string"
                },
                "childIpIds": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "id": {
                    "type": "string"
                },
                "metadata": {
                    "$ref": "#/definitions/betav0.Metadata"
                },
                "metadataResolverAddress": {
                    "type": "string"
                },
                "parentIpIds": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "rootIpIds": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "tokenContract": {
                    "type": "string"
                },
                "tokenId": {
                    "type": "string"
                }
            }
        },
        "betav0.IPAssetResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/betav0.IPAsset"
                }
            }
        },
        "betav0.IPAssetsResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/betav0.IPAsset"
                    }
                }
            }
        },
        "betav0.IpAssetQueryOptions": {
            "type": "object",
            "properties": {
                "orderBy": {
                    "type": "string"
                },
                "orderDirection": {
                    "type": "string"
                },
                "pagination": {
                    "type": "object",
                    "properties": {
                        "limit": {
                            "type": "integer"
                        },
                        "offset": {
                            "type": "integer"
                        }
                    }
                },
                "where": {
                    "type": "object",
                    "properties": {
                        "chainId": {
                            "type": "string"
                        },
                        "metadataResolverAddress": {
                            "type": "string"
                        },
                        "tokenContract": {
                            "type": "string"
                        },
                        "tokenId": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "betav0.IpAssetRequestBody": {
            "type": "object",
            "properties": {
                "options": {
                    "$ref": "#/definitions/betav0.IpAssetQueryOptions"
                }
            }
        },
...
        "betav0.Metadata": {
            "type": "object",
            "properties": {
                "hash": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "registrant": {
                    "type": "string"
                },
                "registrationDate": {
                    "type": "string"
                },
                "uri": {
                    "type": "string"
                }
            }
        },
...
    }
}