# api/internal/handler/betav0/ip_asset.go
package betav0

import (
	"github.com/gin-gonic/gin"
	"github.com/storyprotocol/protocol-api/api/internal/handler/betav0/messages"
	beta_v0 "github.com/storyprotocol/protocol-api/api/internal/models/betav0"
	"github.com/storyprotocol/protocol-api/api/internal/service/thegraph"
	xhttp "github.com/storyprotocol/protocol-api/pkg/http"
	"github.com/storyprotocol/protocol-api/pkg/logger"
	"net/http"
)

// @BasePath /

// GetIPAsset Example godoc
// @Summary Get an IPAsset
// @Schemes
// @Description Retrieve an IPAsset
// @Tags IPAssets
// @Accept json
// @Produce json
// @Param        assetId   path      string  true  "Asset ID"
// @Security ApiKeyAuth
// @param X-API-Key header string true "API Key"
// @Success 200 {object} IPAssetResponse
// @Router /api/v1/assets/{assetId} [get]
func NewGetIPAsset(graphService thegraph.TheGraphServiceBeta, httpClient xhttp.Client) func(c *gin.Context) {
	return func(c *gin.Context) {
		assetId := c.Param("assetId")
		assets, err := graphService.GetIPAsset(assetId)
		if err != nil {
			logger.Errorf("Failed to get asset: %v", err)
			c.JSON(http.StatusInternalServerError, messages.ErrorMessage("Internal server error"))
			return
		}

		c.JSON(http.StatusOK, beta_v0.IPAssetResponse{
			Data: assets,
		})
	}
}

// @BasePath /

// ListIPAssets Example godoc
// @Summary List IPAssets
// @Schemes
// @Description Retrieve a paginated, filtered list of IPAssets
// @Tags IPAssets
// @Accept json
// @Produce json
// @Success 200 {object} IPAssetsResponse
// @Security ApiKeyAuth
// @param X-API-Key header string true "API Key"
// @Param data body betav0.IpAssetRequestBody true "Query Parameters ("where" values are optional. Remove if not using)"
// @Router /api/v1/assets [post]
func NewListIPAssets(graphService thegraph.TheGraphServiceBeta, httpClient xhttp.Client) func(c *gin.Context) {
	return func(c *gin.Context) {
		var requestBody beta_v0.IpAssetRequestBody
		if err := c.BindJSON(&requestBody); err != nil {
			logger.Errorf("Failed to read request body: %v", err)
			requestBody = beta_v0.IpAssetRequestBody{}
		}

		ipAssets, err := graphService.ListIPAssets(fromIPARequestQueryOptions(requestBody.Options))
		if err != nil {
			logger.Errorf("Failed to get registered IP Assets: %v", err)
			c.JSON(http.StatusInternalServerError, messages.ErrorMessage("Internal server error"))
			return
		}

		c.JSON(http.StatusOK, beta_v0.IPAssetsResponse{
			Data: ipAssets,
		})
	}
}

func fromIPARequestQueryOptions(options *beta_v0.IpAssetQueryOptions) *thegraph.TheGraphQueryOptions {
	if options == nil {
		return &thegraph.TheGraphQueryOptions{
			First: 100,
			Skip:  0,
		}
	}

	var queryOptions = &thegraph.TheGraphQueryOptions{}

	if options.Pagination.Limit == 0 {
		options.Pagination.Limit = 100
	}

	queryOptions.First = options.Pagination.Limit
	queryOptions.Skip = options.Pagination.Offset

	queryOptions.Where.MetadataResolverAddress = options.Where.MetadataResolverAddress
	queryOptions.Where.TokenContract = options.Where.TokenContract
	queryOptions.Where.TokenId = options.Where.TokenId
	queryOptions.Where.ChainId = options.Where.ChainId

	return queryOptions
}